Move heartbeat detect from proxy to control panel module
