+++ shardingsphere-control-panel/shardingsphere-cluster/shardingsphere-cluster-heartbeat/pom.xml
@@ -47,5 +47,10 @@
             <artifactId>shardingsphere-cluster-configuration</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.apache.shardingsphere</groupId>
+            <artifactId>shardingsphere-kernel-context</artifactId>
+            <version>${project.version}</version>
+        </dependency>
     </dependencies>
 </project>
